added incremental cache update (INGRID-1672), incremental indexing not implemented

+++ core/src/test/de/ingrid/iplug/csw/dsc/TestUtil.java
@@ -36,14 +36,16 @@ public class TestUtil {
 			return true;
 		}
 		File[] files = directory.listFiles();
-		for (File file : files) {
-			if (file.isDirectory()) {
-				if (!deleteDirectory(file)) {
-					return false;
-				}
-			} else {
-				if (!file.delete()) {
-					return false;
+		if (files != null) {
+			for (File file : files) {
+				if (file.isDirectory()) {
+					if (!deleteDirectory(file)) {
+						return false;
+					}
+				} else {
+					if (!file.delete()) {
+						return false;
+					}
 				}
 			}
 		}
